% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttest.R
\name{cttest}
\alias{cttest}
\title{Common Trend Test}
\usage{
cttest(didfit)
}
\arguments{
\item{didfit}{the object returned by \code{didWD} whose parameters \code{pretime} is \code{TRUE}.}
}
\value{
F statistics testing HO: there is a common trend.
}
\description{
Common Trend Test
}
\examples{
library(plm)
data(stg6, package = 'didWD')
fit <- didWD(stg6, id = 'id', year = 'year', y = 'logy', w = 'w', pretime = TRUE)
cttest(fit)

}
