% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didWD.R
\name{didWD}
\alias{didWD}
\title{DID Regression}
\usage{
didWD(dt, id, year, y, w, wcontinuous = NULL, pretime = FALSE)
}
\arguments{
\item{dt}{A data.frame including \code{id}, \code{year}, \code{y} and \code{w}.}

\item{id}{A coloumn name denotes individuals.}

\item{year}{A coloumn name denotes times. \code{dt$year} is numeric.}

\item{y}{A coloumn name denotes the dependent variable.}

\item{w}{A dummy variable which equals \eqn{D_i\cdot T_i} where \eqn{D_i=1} indicates treated group,
and \eqn{T-i=1} indicates in the treated periods.}

\item{wcontinuous}{A character, the continuous treatment variable's name. Default is \code{NULL}.}

\item{pretime}{Wheather add \eqn{dq\cdot ft, t < q} as independent variables. Default is \code{FALSE}.}
}
\value{
A list including 3 elements, the first is results of regression with Two ways fixed effect,
the second is the data frame \eqn{ft}, and the third is the data frame \eqn{dq}.
}
\description{
Make a DID regression conditional on staggered treatment based on Wooldridge (2021).
}
\examples{
library(plm)
data(stg6, package = 'didWD')
fit <- didWD(stg6, id = 'id', year = 'year', y = 'logy', w = 'w')
lmtest::coeftest(fit$fit, vcov. = vcovHC, method = 'white2')
# aggregate treat effect
xname <- names(coef(fit$fit))
aggeff(fit$fit, xname)

}
