% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyneff.R
\name{dyneff}
\alias{dyneff}
\title{Estimate Dynamic Effect}
\usage{
dyneff(didfit, se = "white2", type = "sss")
}
\arguments{
\item{didfit}{The object returned by \code{didWD} whose parameters \code{pretime} is \code{TRUE}.}

\item{se}{\code{'white1','white2','arellano'}. See the help of \code{plm::vcovHC}.}

\item{type}{\code{"HC0", "sss", "HC1", "HC2", "HC3", "HC4"}. See the help of \code{plm::vcovHC}.}
}
\value{
a list. The 1st element is dynamice effect statistics, and the 2nd element is
independent variables collected.
}
\description{
Estimate Dynamic Effect
}
