% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dq.R
\name{gen_dq}
\alias{gen_dq}
\title{Generate \eqn{d_q}}
\usage{
gen_dq(dt, id, w, stgyr)
}
\arguments{
\item{dt}{A data.frame.}

\item{id}{A index of indivuals.}

\item{w}{A dummy variable which equals \eqn{D_i\cdot T_i} where \eqn{D_i=1} indicates treated group,
and \eqn{T-i=1} indicates in the treated periods.}

\item{stgyr}{A numerica vector about staggered years.}
}
\value{
A data.frame including \eqn{d_q} dummy variables.
}
\description{
Generate \eqn{d_q}
}
\examples{
data(stg6)
gen_dq(stg6, id, w, stgyr = 2013:2016)

}
