% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fst.R
\name{gen_fst}
\alias{gen_fst}
\title{Generate \eqn{fs_t}}
\usage{
gen_fst(dt, tm, stgyr = 2014:2016)
}
\arguments{
\item{dt}{A data.frame.}

\item{tm}{A coloumn name of the time variable.}

\item{stgyr}{A numerica vector about staggered years. For example, \code{2014:2016},
the function would generate \code{f2014,f2015,f2016} three dummy variables.}
}
\value{
A data.frame including time dummy variables.
}
\description{
Generate \eqn{fs_t}
}
\examples{
data(stg6)
gen_fst(stg6, tm = 'year', stgyr = 2014:2016)

}
