% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stg6.R
\docType{data}
\name{stg6}
\alias{stg6}
\title{Example Data}
\format{
A data frame with 3270 rows and 7 varables:
\describe{
\item{id}{individuals ID}
\item{year}{time ID}
\item{y}{the dependent variable}
\item{logy}{\eqn{log(y)}}
\item{w}{A dummy variable which equals \eqn{D_i\cdot T_i} where \eqn{D_i=1} indicates treated group,
and \eqn{T-i=1} indicates in the treated periods.}
\item{x1}{a independent variable}
\item{x2}{a independent variable}
}
}
\usage{
stg6
}
\description{
Example Data
}
\keyword{datasets}
