% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didWD.R
\name{didWD}
\alias{didWD}
\title{DID Regression}
\usage{
didWD(dt, id, year, y, w)
}
\arguments{
\item{dt}{A data.frame including id, year, y and w.}

\item{id}{A coloumn name denotes individuals.}

\item{year}{A coloumn name denotes times.}

\item{y}{A coloumn name denotes the dependent variable.}

\item{w}{A dummy variable which equals \eqn{D_i\cdot T_i} where \eqn{D_i=1} indicates treated group,
and \eqn{T-i=1} indicates in the treated periods.}
}
\value{
Results of Two ways fixed effect
}
\description{
Make a DID regression conditional on staggered treatment based on Wooldridge (2021).
}
\examples{
library(plm)
data(stg6, package = 'didWD')
fit <- didWD(stg6, id = 'id', year = 'year', y = 'logy', w = 'w')
lmtest::coeftest(fit, vcov. = vcovHC, method = 'white2')

}
